imageop_math.c: fix float/int mistake in filter width

+++ src/develop/imageop_math.c
@@ -983,7 +983,7 @@ void dt_iop_clip_and_zoom_mosaic_third_size_xtrans(uint16_t *const out, const ui
   // centered on current sample with edges rounded to nearest input
   // pixel. This seems to be the smallest filter which produces smooth
   // output in most cases.
-  const int kern_width = 3.5f;
+  const float kern_width = 3.5f;
 
 #ifdef _OPENMP
 #pragma omp parallel for default(none) schedule(static)
@@ -993,14 +993,14 @@ void dt_iop_clip_and_zoom_mosaic_third_size_xtrans(uint16_t *const out, const ui
     uint16_t *outc = out + out_stride * y;
 
     const float fy = (y + roi_out->y) * px_footprint;
-    const int miny = MAX(0, (int)(roundf(fy) - kern_width));
-    const int maxy = MIN(roi_in->height-1, (int)(roundf(fy) + kern_width));
+    const int miny = MAX(0, (int)roundf(fy - kern_width));
+    const int maxy = MIN(roi_in->height-1, (int)roundf(fy + kern_width));
 
     float fx = roi_out->x * px_footprint;
     for(int x = 0; x < roi_out->width; x++, fx += px_footprint, outc++)
     {
-      const int minx = MAX(0, (int)(roundf(fx) - kern_width));
-      const int maxx = MIN(roi_in->width-1, (int)(roundf(fx) + kern_width));
+      const int minx = MAX(0, (int)roundf(fx - kern_width));
+      const int maxx = MIN(roi_in->width-1, (int)roundf(fx + kern_width));
 
       const int c = FCxtrans(y, x, roi_out, xtrans);
       int num = 0;
@@ -1024,7 +1024,7 @@ void dt_iop_clip_and_zoom_mosaic_third_size_xtrans_f(float *const out, const flo
                                                      const int32_t in_stride, const uint8_t (*const xtrans)[6])
 {
   const float px_footprint = 1.f / roi_out->scale;
-  const int kern_width = 3.5f;
+  const float kern_width = 3.5f;
 
 #ifdef _OPENMP
 #pragma omp parallel for default(none) schedule(static)
@@ -1034,14 +1034,14 @@ void dt_iop_clip_and_zoom_mosaic_third_size_xtrans_f(float *const out, const flo
     float *outc = out + out_stride * y;
 
     const float fy = (y + roi_out->y) * px_footprint;
-    const int miny = MAX(0, (int)(roundf(fy) - kern_width));
-    const int maxy = MIN(roi_in->height-1, (int)(roundf(fy) + kern_width));
+    const int miny = MAX(0, (int)roundf(fy - kern_width));
+    const int maxy = MIN(roi_in->height-1, (int)roundf(fy + kern_width));
 
     float fx = roi_out->x * px_footprint;
     for(int x = 0; x < roi_out->width; x++, fx += px_footprint, outc++)
     {
-      const int minx = MAX(0, (int)(roundf(fx) - kern_width));
-      const int maxx = MIN(roi_in->width-1, (int)(roundf(fx) + kern_width));
+      const int minx = MAX(0, (int)roundf(fx - kern_width));
+      const int maxx = MIN(roi_in->width-1, (int)roundf(fx + kern_width));
 
       const int c = FCxtrans(y, x, roi_out, xtrans);
       int num = 0;
